refactor: move old tests to brittle
