+++ package.json
@@ -5,7 +5,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "brittle test/websocket.js"
+    "test": "brittle test/*"
   },
   "author": "",
   "license": "ISC",
@@ -18,14 +18,11 @@
     "corestore": "^6.4.2",
     "fastify": "^4.2.0",
     "hyperbee": "^2.0.0",
-    "hypercore": "^10.2.0",
     "hyperswarm": "^4.3.6"
   },
   "devDependencies": {
     "@types/ws": "^8.5.4",
     "brittle": "^3.1.1",
-    "mocha": "^10.0.0",
-    "sinon": "^14.0.0",
     "random-access-memory": "^6.1.0"
   }
 }
